Select the proper matchers for every entry
There are matchers that do not need to run on all entries. The code to
select the matchers based on the entry was repeated in several places,
and missing in other. Move the code into a method to avoid those issues.

The 'header' matchers should run only once. Select the first entry for
this, which should be the request to the main page of the site.

The 'body' matchers should ignore the main-page entry, to avoid false
positives in the content. Make the 'url' matchers ignore the first entry
too, because there should be no need to match against the site URL.