Use a special marker to identify the main request
The first entry, which is the request to the main site, should be used
to match headers, and its content should be ignored.

But if the site redirects to another URL to load the page, then the
request to the main site will not be the first request or the
one with url equals to the main site.

To handle redirections, find the right request following the Location
header, if any. Then mark the request with a special metadata field,
that can be used to identify it.