Handle errors when running JavaScript on the sites
If the Content Security Policy directive is enabled in the site, then
the JavaScript detection function cannot be evaluated, and Splash will
fail with a Lua script error.

But even if the JavaScript matchers cannot be used, the HAR can still be
processed, so let Splash return the HAR normally in case of errors when
evaluating the JavaScript detection function.

The possible error messages are returned as parte of the output
dictionary, in case they are needed.