Make plugins list hints as names instead of results
The list of hints is better as a list names instead of the internal
Result class, which may be created with incomplete data. By using the
plugin names, the proper metadata information can be obtained from the
registered plugin itself.

The from_url field will be set to the main URL requested for detection,
instead of the current entry URL, so rudebot knows where to look at.