Handle Docker errors
Refactor the docker_container context function into a class with
multiple methods in order to organize the code better, due to the
multiple try-catch statements that are required to handle different
problems.

Wrap the Docker exceptions into a custom DockerStartError exception with
simpler error messages, and let them bubble up instead of exiting the
program when requesting the response to the Splash server.