Wait for the Splash container to start
When the Splash container needs to be started, it may take a while to
complete initialization. This may cause the execute Lua script request
to fail, because the sleep between creating the container and sending
the request is fixed to 1 second.

To wait for the container to start, retry a ping to the server until it
succeeds, with increasing sleeps between retries, up to 30 seconds.