Add command line option to list plugins
Add a '--plugins' option to the 'det' command. If set, all the
registered plugins will be printed as a list of names. If '--metadata'
is also set, the plugins will be printed as a list of dictionaries, with
the name, homepage and hints (if any) of each plugin.

Since click still makes the URL argument required, set it with a default
dummy value to ensure the command doesn't fail when '--plugins' is the
only argument given, and print the help manually when a URL is not
passed and detection is intended.

The required flag for the URL argument is to prevent the help printing
the argument as '[URL]' (due to having a default value).