Add support to run detectem with Docker
To run detectem with Docker, add the files Dockerfile-alternate and
docker-compose.yml to the extra/docker directory. The first one is used
to create a local image for detectem and the second one is used to run
detectem as a container with docker-compose, considering the current
Splash dependency.

Also add the instructions to run detectem with Docker to the README.