Print warning when JavaScript matchers failed
Print a warning with the error message returned by Splash when the
JavaScript function failed to match plugins, instead of ignoring the
issue and continuing just with the HAR entries.

The error message returned by Splash seems to be the string
representation of a custom exception object, so in order to show a clear
message it must be parsed (with a regex) and un-escaped.